pyproject.toml
[project]
name = "bot-lor-plus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "aiosqlite>=0.21.0",
    "omegaconf>=2.3.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "parse>=1.20.2",
    "python-decouple>=3.8",
    "sqlalchemy>=2.0.41",
    "tabulate>=0.9.0",
]
